Add Pulse profile side panel
