+++ desktop/src/features/pulse/ui/PulseScreen.tsx
@@ -1,21 +1,45 @@
+import * as React from "react";
+
 import { ChatHeader } from "@/features/chat/ui/ChatHeader";
+import { UserProfilePanel } from "@/features/profile/ui/UserProfilePanel";
 import { PulseView } from "@/features/pulse/ui/PulseView";
 import { useIdentityQuery } from "@/shared/api/hooks";
+import { ProfilePanelProvider } from "@/shared/context/ProfilePanelContext";
+import { useThreadPanelWidth } from "@/shared/hooks/useThreadPanelWidth";
 
 export function PulseScreen() {
   const identityQuery = useIdentityQuery();
+  const [profilePanelPubkey, setProfilePanelPubkey] = React.useState<
+    string | null
+  >(null);
+  const threadPanelWidth = useThreadPanelWidth();
 
   return (
-    <div className="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
-      <ChatHeader
-        description="Notes from people and agents you follow"
-        mode="pulse"
-        overlaysContent
-        title="Pulse"
-      />
+    <ProfilePanelProvider onOpenProfilePanel={setProfilePanelPubkey}>
       <div className="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
-        <PulseView currentPubkey={identityQuery.data?.pubkey} />
+        <ChatHeader
+          description="Notes from people and agents you follow"
+          mode="pulse"
+          overlaysContent
+          title="Pulse"
+        />
+        <div className="flex min-h-0 min-w-0 flex-1 flex-row overflow-hidden">
+          <div className="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
+            <PulseView currentPubkey={identityQuery.data?.pubkey} />
+          </div>
+          {profilePanelPubkey ? (
+            <UserProfilePanel
+              canResetWidth={threadPanelWidth.canReset}
+              currentPubkey={identityQuery.data?.pubkey}
+              onClose={() => setProfilePanelPubkey(null)}
+              onResetWidth={threadPanelWidth.onResetWidth}
+              onResizeStart={threadPanelWidth.onResizeStart}
+              pubkey={profilePanelPubkey}
+              widthPx={threadPanelWidth.widthPx}
+            />
+          ) : null}
+        </div>
       </div>
-    </div>
+    </ProfilePanelProvider>
   );
 }
